requirements.txt
Django==1.6
Pillow==2.2.1
South==0.8.4
argparse==1.2.1
distribute==0.7.3
django-annoying==0.7.7
wsgiref==0.1.2